backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.7.0
pandas==2.2.0
polars==0.20.10
openpyxl==3.1.2
google-generativeai>=0.7.2
instructor>=1.3.3
python-multipart==0.0.9
httpx==0.26.0
